r code is no supporting insertionds better
in insertions we don't have mateid and hence the threshold is too high for them and they are filtered out. so I treat them as they are mateid

in addition, the qual is the sum of quals

and last thing, the filter is the union of the filters. In case there is cohortSize so we give PASS